requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
faiss-cpu>=1.7
umap-learn>=0.5
matplotlib>=3.7
